raftstore: move dynamic config to other mods
