+++ components/test_raftstore/src/node.rs
@@ -11,22 +11,25 @@ use kvproto::raft_serverpb::{self, RaftMessage};
 use raft::eraftpb::MessageType;
 use raft::SnapshotStatus;
 
+use super::*;
 use engine::*;
 use engine_rocks::RocksEngine;
+use tikv::config::ConfigHandler;
 use tikv::config::{ConfigController, TiKvConfig};
 use tikv::import::SSTImporter;
+use tikv::raftstore::coprocessor::config::SplitCheckConfigManager;
 use tikv::raftstore::coprocessor::CoprocessorHost;
 use tikv::raftstore::router::{RaftStoreRouter, ServerRaftStoreRouter};
+use tikv::raftstore::store::config::RaftstoreConfigManager;
+use tikv::raftstore::store::fsm::store::{StoreMeta, PENDING_VOTES_CAP};
 use tikv::raftstore::store::fsm::{RaftBatchSystem, RaftRouter};
 use tikv::raftstore::store::*;
 use tikv::raftstore::Result;
 use tikv::server::Node;
 use tikv::server::Result as ServerResult;
 use tikv_util::collections::{HashMap, HashSet};
-use tikv_util::worker::FutureWorker;
-
-use super::*;
-use tikv::raftstore::store::fsm::store::{StoreMeta, PENDING_VOTES_CAP};
+use tikv_util::config::VersionTrack;
+use tikv_util::worker::{FutureWorker, Worker};
 
 pub struct ChannelTransportCore {
     snap_paths: HashMap<u64, (SnapManager, TempDir)>,
@@ -171,10 +174,12 @@ impl Simulator for NodeCluster {
         let pd_worker = FutureWorker::new("test-pd-worker");
 
         let simulate_trans = SimulateTransport::new(self.trans.clone());
+        let mut raft_store = cfg.raft_store.clone();
+        raft_store.validate().unwrap();
         let mut node = Node::new(
             system,
             &cfg.server,
-            &cfg.raft_store,
+            Arc::new(VersionTrack::new(raft_store)),
             Arc::clone(&self.pd_client),
         );
 
@@ -210,7 +215,32 @@ impl Simulator for NodeCluster {
 
         let store_meta = Arc::new(Mutex::new(StoreMeta::new(PENDING_VOTES_CAP)));
         let local_reader = LocalReader::new(engines.kv.clone(), store_meta.clone(), router.clone());
-        let cfg_controller = ConfigController::new(cfg, Default::default());
+        let mut cfg_controller = ConfigController::new(cfg.clone(), Default::default());
+
+        let mut split_check_worker = Worker::new("split-check");
+        let split_check_runner = SplitCheckRunner::new(
+            Arc::clone(&engines.kv),
+            router.clone(),
+            coprocessor_host.clone(),
+            cfg.coprocessor.clone(),
+        );
+        split_check_worker.start(split_check_runner).unwrap();
+        cfg_controller.register(
+            "coprocessor",
+            Box::new(SplitCheckConfigManager(split_check_worker.scheduler())),
+        );
+
+        let mut raftstore_cfg = cfg.raft_store.clone();
+        raftstore_cfg.validate().unwrap();
+        let raft_store = Arc::new(VersionTrack::new(raftstore_cfg));
+        cfg_controller.register("raft_store", Box::new(RaftstoreConfigManager(raft_store)));
+        let config_client = ConfigHandler::start(
+            cfg.server.advertise_addr,
+            cfg_controller,
+            pd_worker.scheduler(),
+        )
+        .unwrap();
+
         node.start(
             engines.clone(),
             simulate_trans.clone(),
@@ -219,7 +249,8 @@ impl Simulator for NodeCluster {
             store_meta,
             coprocessor_host,
             importer,
-            cfg_controller,
+            split_check_worker,
+            Box::new(config_client) as _,
         )?;
         assert!(engines
             .kv
