+++ src/raftstore/store/fsm/mod.rs
@@ -18,5 +18,5 @@ pub use self::apply::{
 pub use self::peer::{DestroyPeerJob, GroupState, PeerFsm};
 pub use self::store::{
     create_raft_batch_system, new_compaction_listener, RaftBatchSystem, RaftPollerBuilder,
-    RaftRouter, StoreInfo,
+    RaftRouter, StoreInfo, StoreMeta,
 };
