+++ src/raftstore/store/fsm/store.rs
@@ -48,7 +48,7 @@ use crate::raftstore::store::util::is_initial_msg;
 use crate::raftstore::store::worker::{
     CleanupRunner, CleanupSSTRunner, CleanupSSTTask, CleanupTask, CompactRunner, CompactTask,
     ConsistencyCheckRunner, ConsistencyCheckTask, PdRunner, RaftlogGcRunner, RaftlogGcTask,
-    ReadDelegate, RegionRunner, RegionTask, SplitCheckRunner, SplitCheckTask,
+    ReadDelegate, RegionRunner, RegionTask, SplitCheckTask,
 };
 use crate::raftstore::store::DynamicConfig;
 use crate::raftstore::store::PdTask;
@@ -443,7 +443,7 @@ impl<'a, T: Transport, C: PdClient> StoreFsmDelegate<'a, T, C> {
                     self.on_store_unreachable(store_id);
                 }
                 StoreMsg::Start { store } => self.start(store),
-                #[cfg(test)]
+                #[cfg(any(test, feature = "testexport"))]
                 StoreMsg::Validate(f) => f(&self.ctx.cfg),
             }
         }
@@ -986,6 +986,10 @@ impl RaftBatchSystem {
         self.router.clone()
     }
 
+    pub fn apply_router(&self) -> ApplyRouter {
+        self.apply_router.clone()
+    }
+
     // TODO: reduce arguments
     pub fn spawn<T: Transport + 'static, C: PdClient + ConfigClient + 'static>(
         &mut self,
@@ -1061,8 +1065,7 @@ impl RaftBatchSystem {
 
         let engines = builder.engines.clone();
         let snap_mgr = builder.snap_mgr.clone();
-        let cfg = builder.cfg.clone();
-        let cfg = cfg.value();
+        let cfg = builder.cfg.value().clone();
         let store = builder.store.clone();
         let pd_client = builder.pd_client.clone();
         let importer = builder.importer.clone();
@@ -1114,21 +1117,6 @@ impl RaftBatchSystem {
 
         self.apply_system
             .spawn("apply".to_owned(), apply_poller_builder);
-
-        // cfg_controller.register(
-        //     "coprocessor",
-        //     Box::new(SplitCheckConfigManager(
-        //         workers.split_check_worker.scheduler(),
-        //     )),
-        // );
-        let split_check_runner = SplitCheckRunner::new(
-            Arc::clone(&engines.kv),
-            self.router.clone(),
-            workers.coprocessor_host.clone(),
-            // cfg_controller.get_current().coprocessor.clone(),
-            Default::default(),
-        );
-        box_try!(workers.split_check_worker.start(split_check_runner));
 
         let region_runner = RegionRunner::new(
             engines.clone(),
@@ -2120,263 +2108,10 @@ fn is_range_covered<'a, F: Fn(u64) -> &'a metapb::Region>(
 
 #[cfg(test)]
 mod tests {
-    use std::collections::BTreeMap;
-    use std::sync::atomic::AtomicUsize;
-    use std::sync::{mpsc, Arc, Mutex};
-
-    use crate::config::ConfigController;
-    use crate::config::TiKvConfig;
-    use crate::import::SSTImporter;
-    use crate::raftstore::coprocessor::properties::{RangeOffsets, RangeProperties};
-    use crate::raftstore::coprocessor::CoprocessorHost;
-    use crate::raftstore::store::config::RaftstoreConfigManager;
-    use crate::raftstore::store::fsm::*;
-    use crate::raftstore::store::Transport;
-    use engine_rocks::CompactedEvent;
-
-    use batch_system::*;
-    use engine::ALL_CFS;
-    use tempfile::{Builder, TempDir};
-    use tikv_util::collections::HashMap;
-    use tikv_util::config::VersionTrack;
-    use tikv_util::worker::{dummy_scheduler, FutureWorker};
-    use tokio_threadpool;
+    use engine_rocks::RangeOffsets;
+    use engine_rocks::RangeProperties;
 
     use super::*;
-
-    #[derive(Clone)]
-    struct MockTransport;
-    impl Transport for MockTransport {
-        fn send(&mut self, _: RaftMessage) -> Result<()> {
-            unimplemented!()
-        }
-        fn flush(&mut self) {
-            unimplemented!()
-        }
-    }
-
-    struct MockPdClient;
-    impl PdClient for MockPdClient {}
-
-    fn create_tmp_engine(path: &str) -> (TempDir, Engines) {
-        let path = Builder::new().prefix(path).tempdir().unwrap();
-        let db = Arc::new(
-            rocks::util::new_engine(
-                path.path().join("db").to_str().unwrap(),
-                None,
-                ALL_CFS,
-                None,
-            )
-            .unwrap(),
-        );
-        let raft_db = Arc::new(
-            rocks::util::new_engine(path.path().join("raft").to_str().unwrap(), None, &[], None)
-                .unwrap(),
-        );
-        let shared_block_cache = false;
-        (path, Engines::new(db, raft_db, shared_block_cache))
-    }
-
-    fn create_batch_system(
-        cfg: &Config,
-    ) -> (
-        RaftRouter,
-        ApplyRouter,
-        BatchSystem<PeerFsm, StoreFsm>,
-        ApplyBatchSystem,
-    ) {
-        let (store_tx, store_fsm) = StoreFsm::new(cfg);
-        let (apply_router, apply_system) = create_apply_batch_system(&cfg);
-        let (router, system) = create_system(
-            cfg.store_pool_size,
-            cfg.store_max_batch_size,
-            store_tx,
-            store_fsm,
-        );
-        (RaftRouter { router }, apply_router, system, apply_system)
-    }
-
-    fn start_raftstore(
-        cfg: TiKvConfig,
-    ) -> (
-        ConfigController,
-        RaftRouter,
-        ApplyRouter,
-        BatchSystem<PeerFsm, StoreFsm>,
-        ApplyBatchSystem,
-    ) {
-        let (raft_router, apply_router, mut system, mut apply_system) =
-            create_batch_system(&cfg.raft_store);
-        let (_, engines) = create_tmp_engine("store-config");
-        let host = CoprocessorHost::default();
-        let importer = {
-            let dir = Builder::new().prefix("store-config").tempdir().unwrap();
-            Arc::new(SSTImporter::new(dir.path()).unwrap())
-        };
-        let snap_mgr = {
-            let tmp = Builder::new().prefix("store-config").tempdir().unwrap();
-            SnapManager::new(tmp.path().to_str().unwrap(), Some(raft_router.clone()))
-        };
-        let future_poller = tokio_threadpool::Builder::new()
-            .name_prefix("store-config")
-            .pool_size(1)
-            .build()
-            .sender()
-            .clone();
-        let store_meta = Arc::new(Mutex::new(StoreMeta::new(PENDING_VOTES_CAP)));
-        let cfg_track = Arc::new(VersionTrack::new(cfg.raft_store.clone()));
-        let mut cfg_controller = ConfigController::new(cfg, Default::default());
-        cfg_controller.register(
-            "raft_store",
-            Box::new(RaftstoreConfigManager(cfg_track.clone())),
-        );
-        let builder = RaftPollerBuilder {
-            cfg: cfg_track,
-            store: Default::default(),
-            engines,
-            router: raft_router.clone(),
-            split_check_scheduler: dummy_scheduler().0,
-            region_scheduler: dummy_scheduler().0,
-            pd_scheduler: FutureWorker::new("store-config").scheduler(),
-            consistency_check_scheduler: dummy_scheduler().0,
-            cleanup_scheduler: dummy_scheduler().0,
-            raftlog_gc_scheduler: dummy_scheduler().0,
-            apply_router: apply_router.clone(),
-            trans: MockTransport,
-            pd_client: Arc::new(MockPdClient),
-            coprocessor_host: host,
-            importer,
-            snap_mgr,
-            global_stat: Default::default(),
-            store_meta,
-            applying_snap_count: Arc::new(AtomicUsize::new(0)),
-            future_poller,
-        };
-        let apply_poller_builder = ApplyPollerBuilder::new(
-            &builder,
-            ApplyNotifier::Router(raft_router.clone()),
-            apply_router.clone(),
-        );
-        system.spawn("store-config".to_owned(), builder);
-        apply_system.spawn("apply-config".to_owned(), apply_poller_builder);
-        (
-            cfg_controller,
-            raft_router,
-            apply_router,
-            system,
-            apply_system,
-        )
-    }
-
-    fn validate_store<F>(router: &RaftRouter, f: F)
-    where
-        F: FnOnce(&Config) + Send + 'static,
-    {
-        let (tx, rx) = mpsc::channel();
-        router
-            .send_control(StoreMsg::Validate(Box::new(move |cfg: &Config| {
-                f(cfg);
-                tx.send(()).unwrap();
-            })))
-            .unwrap();
-        rx.recv_timeout(Duration::from_secs(3)).unwrap();
-    }
-
-    fn validate_apply<F>(router: &ApplyRouter, region_id: u64, validate: F)
-    where
-        F: FnOnce(bool) + Send + 'static,
-    {
-        let (tx, rx) = mpsc::channel();
-        router.schedule_task(
-            region_id,
-            ApplyTask::Validate(
-                region_id,
-                Box::new(move |(_, sync_log): (_, bool)| {
-                    validate(sync_log);
-                    tx.send(()).unwrap();
-                }),
-            ),
-        );
-        rx.recv_timeout(Duration::from_secs(3)).unwrap();
-    }
-
-    #[test]
-    fn test_update_raftstore_config() {
-        let mut config = TiKvConfig::default();
-        config.validate().unwrap();
-        let (mut cfg_controller, router, _, mut system, mut apply_system) =
-            start_raftstore(config.clone());
-
-        let incoming = config.clone();
-        let raft_store = incoming.raft_store.clone();
-        let rollback = cfg_controller.update_or_rollback(incoming).unwrap();
-
-        // config should not change
-        assert_eq!(rollback.right(), Some(false));
-        validate_store(&router, move |cfg: &Config| {
-            assert_eq!(cfg, &raft_store);
-        });
-
-        // dispatch updated config
-        let mut raft_store = config.raft_store.clone();
-        raft_store.messages_per_tick = 12345;
-        raft_store.raft_log_gc_threshold = 54321;
-        let mut incoming = config;
-        incoming.raft_store = raft_store.clone();
-        let rollback = cfg_controller.update_or_rollback(incoming).unwrap();
-
-        // config should be updated
-        assert_eq!(rollback.right(), Some(true));
-        validate_store(&router, move |cfg: &Config| {
-            assert_eq!(cfg, &raft_store);
-        });
-
-        apply_system.shutdown();
-        system.shutdown();
-    }
-
-    #[test]
-    fn test_update_apply_store_config() {
-        let mut config = TiKvConfig::default();
-        config.raft_store.sync_log = true;
-        config.validate().unwrap();
-        let (mut cfg_controller, raft_router, apply_router, mut system, mut apply_system) =
-            start_raftstore(config.clone());
-
-        // register region
-        let region_id = 1;
-        let mut reg = Registration::default();
-        reg.region.set_id(region_id);
-        apply_router.schedule_task(region_id, ApplyTask::Registration(reg));
-
-        let rollback = cfg_controller.update_or_rollback(config.clone()).unwrap();
-
-        // config should not change
-        assert_eq!(rollback.right(), Some(false));
-        validate_store(&raft_router, move |cfg: &Config| {
-            assert_eq!(cfg.sync_log, true);
-        });
-        validate_apply(&apply_router, region_id, |sync_log| {
-            assert_eq!(sync_log, true);
-        });
-
-        // dispatch updated config
-        let mut incoming = config;
-        incoming.raft_store.sync_log = false;
-        let rollback = cfg_controller.update_or_rollback(incoming).unwrap();
-
-        // both configs should be updated
-        assert_eq!(rollback.right(), Some(true));
-        validate_store(&raft_router, move |cfg: &Config| {
-            assert_eq!(cfg.sync_log, false);
-        });
-        validate_apply(&apply_router, region_id, |sync_log| {
-            assert_eq!(sync_log, false);
-        });
-
-        apply_system.shutdown();
-        system.shutdown();
-    }
 
     #[test]
     fn test_calc_region_declined_bytes() {
