+++ src/raftstore/store/msg.rs
@@ -366,7 +366,7 @@ pub enum StoreMsg {
     },
 
     /// Messge only used for test
-    #[cfg(test)]
+    #[cfg(any(test, feature = "testexport"))]
     Validate(Box<dyn FnOnce(&crate::raftstore::store::Config) + Send>),
 }
 
@@ -390,7 +390,7 @@ impl fmt::Debug for StoreMsg {
             ),
             StoreMsg::Tick(tick) => write!(fmt, "StoreTick {:?}", tick),
             StoreMsg::Start { ref store } => write!(fmt, "Start store {:?}", store),
-            #[cfg(test)]
+            #[cfg(any(test, feature = "testexport"))]
             StoreMsg::Validate(_) => write!(fmt, "Validate config"),
         }
     }
