+++ tests/integrations/config/mod.rs
@@ -22,6 +22,7 @@ use tikv::storage::config::{BlockCacheConfig, Config as StorageConfig};
 use tikv_util::config::{ReadableDuration, ReadableSize};
 use tikv_util::security::SecurityConfig;
 
+mod dynamic;
 mod test_config_client;
 
 #[test]
