+++ tests/integrations/raftstore/test_bootstrap.rs
@@ -11,12 +11,14 @@ use kvproto::raft_serverpb::RegionLocalState;
 use engine::*;
 use test_raftstore::*;
 use tikv::config::ConfigController;
+use tikv::config::ConfigHandler;
 use tikv::import::SSTImporter;
 use tikv::raftstore::coprocessor::CoprocessorHost;
 use tikv::raftstore::store::fsm::store::StoreMeta;
 use tikv::raftstore::store::{bootstrap_store, fsm, SnapManager};
 use tikv::server::Node;
-use tikv_util::worker::FutureWorker;
+use tikv_util::config::VersionTrack;
+use tikv_util::worker::{FutureWorker, Worker};
 
 fn test_bootstrap_idempotent<T: Simulator>(cluster: &mut Cluster<T>) {
     // assume that there is a node  bootstrap the cluster and add region in pd successfully
@@ -56,7 +58,12 @@ fn test_node_bootstrap_with_prepared_data() {
     );
     let tmp_mgr = Builder::new().prefix("test_cluster").tempdir().unwrap();
 
-    let mut node = Node::new(system, &cfg.server, &cfg.raft_store, Arc::clone(&pd_client));
+    let mut node = Node::new(
+        system,
+        &cfg.server,
+        Arc::new(VersionTrack::new(cfg.raft_store.clone())),
+        Arc::clone(&pd_client),
+    );
     let snap_mgr = SnapManager::new(tmp_mgr.path().to_str().unwrap(), Some(node.get_router()));
     let pd_worker = FutureWorker::new("test-pd-worker");
 
@@ -85,8 +92,13 @@ fn test_node_bootstrap_with_prepared_data() {
         Arc::new(SSTImporter::new(dir).unwrap())
     };
 
-    let cfg_controller = ConfigController::new(cfg, Default::default());
-
+    let cfg_controller = ConfigController::new(cfg.clone(), Default::default());
+    let config_client = ConfigHandler::start(
+        cfg.server.advertise_addr,
+        cfg_controller,
+        pd_worker.scheduler(),
+    )
+    .unwrap();
     // try to restart this node, will clear the prepare data
     node.start(
         engines,
@@ -96,7 +108,8 @@ fn test_node_bootstrap_with_prepared_data() {
         Arc::new(Mutex::new(StoreMeta::new(0))),
         coprocessor_host,
         importer,
-        cfg_controller,
+        Worker::new("split"),
+        Box::new(config_client),
     )
     .unwrap();
     assert!(Arc::clone(&engine)
